+++ biobb_dna/__init__.py
@@ -1,4 +1,4 @@
 name = "biobb_dna"
 __all__ = ["dna", "curvesplus", "backbone", "stiffness",
            "interbp_correlations", "intrabp_correlations"]
-__version__ = "4.2.2"
+__version__ = "4.2.3"