+++ biobb_dna/curvesplus/canal_unzip.py
@@ -27,6 +27,7 @@ class CanalUnzip(BiobbObject):
             * **correlation** (*str*) - (None) Correlation indexes separated by underscore (ie '98_165'), only for 'corr' type.
             * **remove_tmp** (*bool*) - (True) [WF property] Remove temporal files.
             * **restart** (*bool*) - (False) [WF property] Do not execute if output files exist.
+            * **sandbox_path** (*str*) - ("./") [WF property] Parent path to the sandbox directory.
     Examples:
         This is a use example of how to use the building block from Python::