+++ biobb_dna/interbp_correlations/interhpcorr.py
@@ -33,7 +33,6 @@ class InterHelParCorrelation(BiobbObject):
             * **remove_tmp** (*bool*) - (True) [WF property] Remove temporal files.
             * **restart** (*bool*) - (False) [WF property] Do not execute if output files exist.
             * **sandbox_path** (*str*) - ("./") [WF property] Parent path to the sandbox directory.
-            * **basepair** (*str*) - (None) Name of basepair analyzed.
 
     Examples:
         This is a use example of how to use the building block from Python::