+++ biobb_dna/intrabp_correlations/intrahpcorr.py
@@ -33,7 +33,6 @@ class IntraHelParCorrelation(BiobbObject):
             * **remove_tmp** (*bool*) - (True) [WF property] Remove temporal files.
             * **restart** (*bool*) - (False) [WF property] Do not execute if output files exist.
             * **sandbox_path** (*str*) - ("./") [WF property] Parent path to the sandbox directory.
-            * **base** (*str*) - (None) Name of base analyzed.
 
     Examples:
         This is a use example of how to use the building block from Python::